requirements.txt
asgiref==3.3.1
autopep8==1.5.7
certifi==2020.11.8
cloudinary==1.23.0
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==14.2.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-registration==3.1.1
django-registration-redux==2.9
djangorestframework==3.12.4
gunicorn==19.9.0
importlib-metadata==1.7.0
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.7.0
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
